+++ packages/extension-base/src/services/transaction-service/constants.ts
@@ -1,4 +1,4 @@
 // Copyright 2019-2022 @subwallet/extension-base authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export const TRANSACTION_TIMEOUT = 2 * 60 * 1000;
+export const TRANSACTION_TIMEOUT = 5 * 60 * 1000;