+++ packages/extension-koni-ui/src/assets/logo/index.ts
@@ -15,7 +15,8 @@ export const DefaultLogosMap: Record<string, string> = {
   polkadot_vault: './images/projects/polkadot-vault.png',
   walletconnect: './images/projects/walletconnect.png',
   banxa: './images/projects/banxa.png',
-  coinbase: './images/projects/coinbase.png'
+  coinbase: './images/projects/coinbase.png',
+  stellaswap: './images/projects/stellaswap.png'
 };
 
 export const IconMap = {