+++ packages/extension-koni-ui/src/components/Modal/Earning/EarningInstructionModal.tsx
@@ -1,22 +1,23 @@
 // Copyright 2019-2022 @subwallet/extension-koni-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import { NotificationType } from '@subwallet/extension-base/background/KoniTypes';
 import { getValidatorLabel } from '@subwallet/extension-base/koni/api/staking/bonding/utils';
 import { _STAKING_CHAIN_GROUP } from '@subwallet/extension-base/services/earning-service/constants';
 import { calculateReward } from '@subwallet/extension-base/services/earning-service/utils';
 import { YieldPoolType } from '@subwallet/extension-base/types';
 import { balanceFormatter, detectTranslate, formatNumber } from '@subwallet/extension-base/utils';
-import InstructionItem from '@subwallet/extension-koni-ui/components/Common/InstructionItem';
+import { InstructionItem } from '@subwallet/extension-koni-ui/components';
 import { getInputValuesFromString } from '@subwallet/extension-koni-ui/components/Field/AmountInput';
 import { EARNING_DATA_RAW, EARNING_INSTRUCTION_MODAL } from '@subwallet/extension-koni-ui/constants';
 import { useSelector } from '@subwallet/extension-koni-ui/hooks';
 import { earlyValidateJoin } from '@subwallet/extension-koni-ui/messaging';
 import { AlertDialogProps, PhosphorIcon, ThemeProps } from '@subwallet/extension-koni-ui/types';
 import { getBannerButtonIcon } from '@subwallet/extension-koni-ui/utils';
-import { BackgroundIcon, Button, Icon, Logo, ModalContext, SwModal, Tag } from '@subwallet/react-ui';
+import { BackgroundIcon, Button, Icon, ModalContext, SwModal } from '@subwallet/react-ui';
 import { getAlphaColor } from '@subwallet/react-ui/lib/theme/themes/default/colorAlgorithm';
 import CN from 'classnames';
-import { CaretDown, PlusCircle } from 'phosphor-react';
+import { CaretCircleLeft, CaretDown, CheckCircle, PlusCircle, XCircle } from 'phosphor-react';
 import React, { useCallback, useContext, useEffect, useMemo, useRef, useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import styled from 'styled-components';
@@ -53,6 +54,7 @@ const Component: React.FC<Props> = (props: Props) => {
   const { currentAccount } = useSelector((state) => state.accountState);
   const [loading, setLoading] = useState(false);
   const [isScrollEnd, setIsScrollEnd] = useState(false);
+  const [isDisableEarnButton, setDisableEarnButton] = useState(true);
   const poolInfo = useMemo(() => poolInfoMap[slug], [poolInfoMap, slug]);
   const title = useMemo(() => {
     if (!poolInfo) {
@@ -122,11 +124,7 @@ const Component: React.FC<Props> = (props: Props) => {
     }
 
     if (shortName) {
-      if (shortName === 'Stellaswap') {
-        result = result.replace('{{shortName}}', 'StellaSwap');
-      } else {
-        result = result.replace('{{shortName}}', shortName);
-      }
+      result = result.replace('{{shortName}}', shortName);
     }
 
     return result;
@@ -148,34 +146,6 @@ const Component: React.FC<Props> = (props: Props) => {
         return 'Supply to earn';
     }
   }, [poolInfo]);
-
-  const tags = useMemo(() => {
-    const asset = assetRegistry[poolInfo.metadata.inputAsset];
-    const symbol = asset.symbol;
-
-    if (poolInfo.statistic && 'assetEarning' in poolInfo.statistic && poolInfo.statistic?.assetEarning) {
-      const assetEarning = poolInfo.statistic?.assetEarning;
-      const data = assetEarning.map((item) => {
-        const result: { slug: string; apy: number; symbol: string } = { slug: item.slug, apy: 0, symbol: symbol };
-
-        result.slug = item.slug;
-
-        if (!item.apy) {
-          const rs = calculateReward(item?.apr || 0);
-
-          result.apy = rs.apy || 0;
-        } else {
-          result.apy = item.apy;
-        }
-
-        return result;
-      });
-
-      return data.filter((item) => item.apy);
-    }
-
-    return [];
-  }, [assetRegistry, poolInfo.metadata.inputAsset, poolInfo.statistic]);
 
   const setVisible = useCallback(
     (show: boolean) => {
@@ -230,8 +200,8 @@ const Component: React.FC<Props> = (props: Props) => {
     switch (poolInfo.type) {
       case YieldPoolType.NOMINATION_POOL: {
         const _label = getValidatorLabel(poolInfo.chain);
-        const label = _label.slice(0, 1).toLowerCase().concat(_label.slice(1)).concat('s');
         const maxCandidatePerFarmer = poolInfo.statistic?.maxCandidatePerFarmer || 0;
+        const label = `${_label.charAt(0).toLowerCase() + _label.substr(1)}${maxCandidatePerFarmer > 1 ? 's' : ''}`;
         const inputAsset = assetRegistry[poolInfo.metadata.inputAsset];
         const maintainAsset = assetRegistry[poolInfo.metadata.maintainAsset];
         const paidOut = poolInfo.statistic?.eraTime;
@@ -254,6 +224,7 @@ const Component: React.FC<Props> = (props: Props) => {
 
             if (paidOut !== undefined) {
               replaceEarningValue(_item, '{paidOut}', paidOut.toString());
+              replaceEarningValue(_item, '{paidOutTimeUnit}', paidOut > 1 ? 'hours' : 'hour');
             }
 
             return _item;
@@ -265,8 +236,8 @@ const Component: React.FC<Props> = (props: Props) => {
 
       case YieldPoolType.NATIVE_STAKING: {
         const _label = getValidatorLabel(poolInfo.chain);
-        const label = _label.slice(0, 1).toLowerCase().concat(_label.slice(1)).concat('s');
         const maxCandidatePerFarmer = poolInfo.statistic?.maxCandidatePerFarmer || 0;
+        const label = `${_label.charAt(0).toLowerCase() + _label.substr(1)}${maxCandidatePerFarmer > 1 ? 's' : ''}`;
         const inputAsset = assetRegistry[poolInfo.metadata.inputAsset];
         const maintainAsset = assetRegistry[poolInfo.metadata.maintainAsset];
         const paidOut = poolInfo.statistic?.eraTime;
@@ -283,12 +254,14 @@ const Component: React.FC<Props> = (props: Props) => {
               const _item: BoxProps = { ...item, id: item.icon, icon: getBannerButtonIcon(item.icon) as PhosphorIcon };
 
               replaceEarningValue(_item, '{validatorNumber}', maxCandidatePerFarmer.toString());
+              replaceEarningValue(_item, '{dAppString}', maxCandidatePerFarmer > 1 ? 'dApps' : 'dApp');
               replaceEarningValue(_item, '{periodNumb}', unBondedTime);
               replaceEarningValue(_item, '{maintainBalance}', maintainBalance);
               replaceEarningValue(_item, '{maintainSymbol}', maintainSymbol);
 
               if (paidOut !== undefined) {
                 replaceEarningValue(_item, '{paidOut}', paidOut.toString());
+                replaceEarningValue(_item, '{paidOutTimeUnit}', paidOut > 1 ? 'hours' : 'hour');
               }
 
               return _item;
@@ -306,6 +279,7 @@ const Component: React.FC<Props> = (props: Props) => {
 
             if (paidOut !== undefined) {
               replaceEarningValue(_item, '{paidOut}', paidOut.toString());
+              replaceEarningValue(_item, '{paidOutTimeUnit}', paidOut > 1 ? 'hours' : 'hour');
             }
 
             return _item;
@@ -390,6 +364,11 @@ const Component: React.FC<Props> = (props: Props) => {
 
       case 'Bifrost Polkadot': {
         urlParam = '#bifrost';
+
+        if (poolInfo.slug === 'MANTA___liquid_staking___bifrost_dot') {
+          urlParam = '#vmanta-on-bifrost';
+        }
+
         break;
       }
 
@@ -403,6 +382,11 @@ const Component: React.FC<Props> = (props: Props) => {
         break;
       }
 
+      case 'Parallel': {
+        urlParam = '#parallel';
+        break;
+      }
+
       case 'Stellaswap': {
         urlParam = '#stellaswap';
         break;
@@ -412,7 +396,7 @@ const Component: React.FC<Props> = (props: Props) => {
     const url = `https://docs.subwallet.app/main/web-dashboard-user-guide/earning/faqs${urlParam}`;
 
     open(url);
-  }, [poolInfo.metadata.shortName]);
+  }, [poolInfo.metadata.shortName, poolInfo.slug]);
 
   const onClickButton = useCallback(() => {
     const time = Date.now();
@@ -427,12 +411,12 @@ const Component: React.FC<Props> = (props: Props) => {
     const onError = (message: string) => {
       openAlert({
         title: t('Pay attention!'),
+        type: NotificationType.ERROR,
         content: message,
         okButton: {
           text: t('I understand'),
-          onClick: () => {
-            closeAlert();
-          }
+          onClick: closeAlert,
+          icon: CheckCircle
         }
       });
     };
@@ -478,10 +462,8 @@ const Component: React.FC<Props> = (props: Props) => {
       return;
     }
 
-    if (!isScrollEnd && scrollRef.current.scrollTop >= scrollRef.current.scrollHeight - 500) {
-      setIsScrollEnd(true);
-    }
-  }, [isScrollEnd]);
+    setIsScrollEnd(scrollRef.current.scrollTop >= scrollRef.current.scrollHeight - 500);
+  }, []);
 
   const closeModal = useCallback(() => {
     setVisible(false);
@@ -495,40 +477,75 @@ const Component: React.FC<Props> = (props: Props) => {
     }
   }, [inactiveModal, poolInfo]);
 
+  useEffect(() => {
+    if (isScrollEnd) {
+      setDisableEarnButton(false);
+    }
+  }, [isScrollEnd]);
+
   const footerNode = (
     <>
-      <div className='__footer-more-information-text'>
-        <Trans
-          components={{
-            highlight: (
-              <a
-                className='__link'
-                onClick={onClickFaq}
-                rel='noopener noreferrer'
-                target='_blank'
-              />
-            )
-          }}
-          i18nKey={detectTranslate('Scroll down to continue. For more information and staking instructions, read <highlight>this FAQ</highlight>')}
+      <div className='__footer-more-information-text-wrapper'>
+        <div className='__footer-more-information-text'>
+          <Trans
+            components={{
+              highlight: (
+                <a
+                  className='__link'
+                  onClick={onClickFaq}
+                  rel='noopener noreferrer'
+                  target='_blank'
+                />
+              )
+            }}
+            i18nKey={detectTranslate('For more information and staking instructions, read <highlight>this FAQ</highlight>')}
+          />
+        </div>
+
+        <Button
+          className={'__scroll-to-end-button'}
+          disabled={isScrollEnd}
+          icon={<Icon phosphorIcon={CaretDown} />}
+          onClick={onScrollContent}
+          shape={'circle'}
+          size={'xs'}
         />
       </div>
-      {isShowStakeMoreButton && (
+
+      <div className={'__buttons'}>
         <Button
           block={true}
-          className={'__stake-more-button'}
-          disabled={!isScrollEnd}
+          className={'__cancel-button'}
           icon={
             <Icon
-              phosphorIcon={PlusCircle}
+              phosphorIcon={isShowStakeMoreButton ? CaretCircleLeft : XCircle}
               weight='fill'
             />
           }
-          loading={loading}
-          onClick={onClickButton}
+          onClick={closeModal}
+          schema={'secondary'}
         >
-          {buttonTitle}
+          {isShowStakeMoreButton ? t('Back') : t('Close')}
         </Button>
-      )}
+
+        {isShowStakeMoreButton && (
+          <Button
+            block={true}
+            className={'__stake-more-button'}
+            disabled={isDisableEarnButton}
+            icon={
+              <Icon
+                phosphorIcon={PlusCircle}
+                weight='fill'
+              />
+            }
+            loading={loading}
+            onClick={onClickButton}
+          >
+            {buttonTitle}
+          </Button>
+        )}
+      </div>
     </>
   );
 
@@ -539,44 +556,18 @@ const Component: React.FC<Props> = (props: Props) => {
   return (
     <SwModal
       className={CN(className)}
+      closable={false}
       destroyOnClose={true}
       footer={footerNode}
       id={modalId}
       onCancel={closeModal}
-      title={t('Instruction')}
+      title={title}
     >
       <div
         className={'__scroll-container'}
         onScroll={onScrollToAcceptButton}
         ref={scrollRef}
       >
-        <div className={'__main-title-area'}>
-          <div className={'__main-title'}>{title}</div>
-
-          {!!(tags && tags.length) && (
-            <div
-              className={'__tags-container'}
-            >
-              {tags.map(({ apy, slug: tagSlug, symbol }) => (
-                <Tag
-                  bgType={'gray'}
-                  className={'__tag-item'}
-                  icon={(
-                    <Logo
-                      size={16}
-                      token={tagSlug.toLowerCase()}
-                    />
-                  )}
-                  key={tagSlug}
-                  shape={'round'}
-                >
-                  {`${formatNumber(apy, 0, balanceFormatter)}% ${symbol}`}
-                </Tag>
-              ))}
-            </div>
-          )}
-        </div>
-
         <div className={'__instruction-items-container'}>
           {data.map((item, index) => {
             return (
@@ -601,28 +592,19 @@ const Component: React.FC<Props> = (props: Props) => {
           })}
         </div>
       </div>
-
-      <div className={'__scroll-to-end-button-wrapper'}>
-        {!isScrollEnd && <Button
-          className={'__scroll-to-end-button'}
-          icon={<Icon phosphorIcon={CaretDown} />}
-          onClick={onScrollContent}
-          shape={'circle'}
-          size={'xs'}
-        />}
-      </div>
     </SwModal>
   );
 };
 
 const EarningInstructionModal = styled(Component)<Props>(({ theme: { token } }: Props) => {
   return ({
-    height: '100%',
+    '.ant-sw-modal-content': {
+      maxHeight: 584,
+      height: 584
+    },
 
     '.ant-sw-modal-content.ant-sw-modal-content': {
-      maxHeight: 'none',
-      height: '100%',
-      borderRadius: 0
+
     },
 
     '.ant-sw-modal-body.ant-sw-modal-body': {
@@ -633,28 +615,21 @@ const EarningInstructionModal = styled(Component)<Props>(({ theme: { token } }: 
       borderTop: 0
     },
 
-    '.__main-title': {
-      fontSize: token.fontSizeHeading4,
-      lineHeight: token.lineHeightHeading4,
-      color: token.colorTextLight1,
-      fontWeight: token.headingFontWeight,
-      textAlign: 'center'
+    '.__footer-more-information-text-wrapper': {
+      display: 'flex',
+      gap: token.size
     },
 
     '.__footer-more-information-text': {
       color: token.colorTextLight4,
       fontSize: token.fontSize,
       lineHeight: token.lineHeight,
-      textAlign: 'center',
+      textAlign: 'left',
 
       '.__link': {
         textDecoration: 'underline',
         color: token.colorPrimary
       }
-    },
-
-    '.__stake-more-button': {
-      marginTop: token.margin
     },
 
     '.__instruction-item': {
@@ -677,38 +652,21 @@ const EarningInstructionModal = styled(Component)<Props>(({ theme: { token } }: 
       scrollBehavior: 'smooth'
     },
 
-    '.__main-title-area': {
-      marginBottom: token.margin
+    '.ant-sw-sub-header-title-content': {
+      'white-space': 'normal'
     },
 
-    '.__tags-container': {
+    '.ant-sw-header-container-center .ant-sw-header-center-part': {
+      position: 'relative',
+      maxWidth: 284,
+      marginLeft: 'auto',
+      marginRight: 'auto'
+    },
+
+    '.__buttons': {
+      marginTop: token.margin,
       display: 'flex',
-      flexWrap: 'wrap',
-      justifyContent: 'center',
-      gap: token.sizeXS,
-      marginTop: token.marginXS,
-
-      '.ant-image-img, .ant-logo, .ant-image': {
-        display: 'block'
-      }
-    },
-
-    '.__tag-item': {
-      display: 'inline-flex',
-      alignItems: 'center',
-      marginRight: 0,
-      gap: token.sizeXXS,
-      color: token.colorSecondary
-    },
-
-    '.__scroll-to-end-button-wrapper': {
-      position: 'relative'
-    },
-
-    '.__scroll-to-end-button': {
-      position: 'absolute',
-      bottom: token.sizeXXS,
-      right: token.size
+      gap: token.sizeXXS
     }
   });
 });